Mod managment polish
Elevated quiet and m28ai to
official mods.

fixed quiet datapath so
git-quiet will work with it
instead of looking in the wrong place.

setup tracking for m28 so that
we auto-enable it on startup if it's
detected.  This behavior can currently
be overriden per-session (e.g. for
debugging) by disabling m28,
but by default it's enabled globally.

This allows us to keep pace with
the m28 release model whereby
the uid is bumped instead of held
constant.  Elegant enough solution
for now.